refactor: transform exercise
