--- problems/transform/aliens.json
@@ -1,321 +1,0 @@
-["Abaddon",
-"Abzorbaloff",
-"Adipose",
-"Aeolian",
-"Aggedor",
-"Akhaten humanoid",
-"Alpha Centauran",
-"Alzarian",
-"Alzarian spider",
-"Ambassadors",
-"Ancient Lights",
-"Androgum",
-"Andromedan",
-"Androzani bat",
-"Androzani tree",
-"Anethan",
-"Animus",
-"Anti-matter organism",
-"Anubian",
-"Arcateenian",
-"Arcturan",
-"Argolin",
-"Aridian",
-"Atraxi",
-"Auton",
-"Axos/Axon/Axonite",
-"Bandril",
-"Bane",
-"Bell Plant",
-"Berserker",
-"Biomechanoid",
-"Blowfish",
-"Bodach",
-"Brain parasite",
-"Carrionite",
-"Cash Cow",
-"Cassandra O'Brien.∆17",
-"Castrovalva humanoid",
-"Catkind",
-"Celestial Toymaker",
-"Cell 114",
-"Centuripede",
-"Chameleon",
-"Cheetah person",
-"Chelonian",
-"Chimeron",
-"Chloris humanoid",
-"Chronovore",
-"Chula",
-"Cowled figure",
-"Crespallion",
-"Crinothian",
-"Crooked person",
-"Cryon",
-"Cyberman",
-"Cyberman",
-"Cybermat",
-"Cybershade",
-"Dæmon",
-"Dalek",
-"Dalek/Human hybrid",
-"Dalek puppet",
-"Dark Hoarde",
-"Dauntless Prison inmates",
-"Diblosian",
-"Dido humanoid",
-"Digi-human",
-"Dogon",
-"Dominator",
-"Draconian",
-"Drahvin",
-"Drashig",
-"Drornidian",
-"Dulcian",
-"Duroc",
-"Eight Legs",
-"Eknodine",
-"Elder",
-"Empty Child plague",
-"Erasmus Darkening",
-"Eternal",
-"Etydion",
-"Eve",
-"Exxilon",
-"Face of Boe",
-"Fairy",
-"Family of Blood",
-"Fear entity",
-"Fendahl",
-"Fenric",
-"Fleshkind",
-"Flying stingray",
-"Foamasi",
-"Force-grown clone",
-"Futurekind",
-"Ganger",
-"Gastropod",
-"Gaztak",
-"Gee-jee fly",
-"Gelth",
-"Gond",
-"Gorgon",
-"Graske",
-"Great Intelligence",
-"Great Vampire",
-"Groske",
-"Gryffen family ghosts",
-"Guardian of Time",
-"Gubbage cone",
-"Haemo-Goth",
-"Haemovore",
-"Hand of Omega",
-"Hath",
-"Headless monk",
-"Hetocumtek",
-"Hitchhiker",
-"Hooloovoo",
-"Ice Warrior",
-"Inter Minorian",
-"International Gallery paintings",
-"Isolus",
-"Jadondan",
-"Jagaroth",
-"Jagrafess",
-"Jeggorabax energy entity",
-"Jixen",
-"Judoon",
-"Kahler",
-"Kaled",
-"Karfelon",
-"Karn humanoid",
-"Kastrian",
-"Keller Machine",
-"Kinda",
-"Kitling",
-"Korven",
-"Kraal",
-"Krafayis",
-"Krarg",
-"Krillitane",
-"Kroll",
-"Krontep humanoid",
-"Kroton",
-"Krynoid",
-"Lakertyan",
-"Land of Fiction beings",
-"Levithian",
-"Logopolitan",
-"Lucanian",
-"Lugal-Irra-Kush",
-"Lurman",
-"Macra",
-"Magma beast",
-"Malmooth",
-"Malus",
-"Mandragora Helix",
-"Mandrel",
-"Manussan",
-"Mara",
-"Marshman",
-"Master Brain",
-"Mayfly",
-"Mede",
-"Megara",
-"Melissa Majoria bees",
-"Memory worm",
-"Mentor",
-"Metalkind",
-"Minotaur",
-"Minyan",
-"Mire beast",
-"Mogarian",
-"Monoid",
-"Morestran",
-"Morlox",
-"Morok",
-"Mutated maggots/flies",
-"Mutated rat",
-"Mute",
-"Myrka",
-"Navarino",
-"Nestene Consciousness",
-"New Human",
-"Night Travellers",
-"Nimon",
-"Nostrovite",
-"Ogri",
-"Ogron",
-"Ood",
-"Ood Brain",
-"Oroborus",
-"Osirian",
-"Pan-babylonian",
-"Patchwork person",
-"Peg doll",
-"Peladonian",
-"Pelushi",
-"Pipe person",
-"Plasmaton",
-"Plasmavore",
-"''Platform One'' guests",
-"Posicarian",
-"Primord",
-"Proamonian",
-"Proto-human",
-"Psychic Circus",
-"Psychic pollen",
-"Pyrovile",
-"Qetesh",
-"Racnoss",
-"Raxacoricofallapatorian",
-"Reaper",
-"Red Leech",
-"Refusian",
-"Ribosian",
-"Richard Lazarus",
-"Rill",
-"Rutan",
-"Sand beast",
-"Saturnyn",
-"Scarecrow",
-"Screamer",
-"Sea Devil",
-"Seaweed creature",
-"Seed pod",
-"Sensorite",
-"Serpentine defence mechanism",
-"Sex gas",
-"Shadow Proclamation humanoids",
-"Shakri",
-"Shansheeth",
-"Shrivenzale",
-"Silurian",
-"Siren",
-"Sirian",
-"Skarasen",
-"Skonnan",
-"Skullion",
-"Sky fish",
-"Slyther",
-"Solonian",
-"Sontaran",
-"Space Pig",
-"Spiridon",
-"Star Whale",
-"Stigorax",
-"Stingray",
-"Swampie",
-"Sycorax",
-"Taran",
-"Tenza",
-"Terileptil",
-"Terraberserker",
-"Tetrap",
-"Thal",
-"Tharil",
-"The 4-5-6",
-"The Beast",
-"The Captain",
-"The Destroyer",
-"The Ergon",
-"The Flood",
-"The Garm",
-"The Hunger",
-"The Raak",
-"The Shadow",
-"The Silence",
-"The Swarm",
-"The Trickster",
-"The Wire",
-"Thoros Alphan",
-"Tigellan",
-"Time Beetle",
-"Time Brain",
-"Time Lord",
-"Tivoli species",
-"Toclafane",
-"Torajii",
-"Tractator",
-"Trakenite",
-"Travist Polong",
-"Trion",
-"Tritovore",
-"Tumour alien",
-"Tythonian",
-"Ukkan",
-"Ultramancer",
-"Urbankan",
-"Usurian",
-"Uvodni",
-"Uxariean",
-"Validium",
-"Vardan",
-"Varga plant",
-"Vashta Nerada",
-"Veil",
-"Vervoid",
-"Vespiform",
-"Vigil",
-"Vinvocci",
-"Vishklar",
-"Visian",
-"Vogan",
-"Voord",
-"Vortis",
-"War Lord",
-"Weeping Angel",
-"Weevil",
-"Werewolf",
-"Whisper men",
-"Winder",
-"Wirrn",
-"Wolfweed",
-"Wood beast",
-"Xeraphin",
-"Xeron",
-"Xylok",
-"Zanak humanoid",
-"Zocci",
-"Zolfa-Thuran",
-"Zygon"]
