+++ problems/transform/index.js
@@ -1,59 +1,13 @@
-var fs = require('fs')
-var path = require('path')
-var verify = require('adventure-verify')
-var concat = require('concat-stream')
-var spawn = require('child_process').spawn
-var aliens = require('./aliens.json')
+const path = require('path')
+const comparestdout = require('workshopper-exercise/comparestdout')
 
-exports.problem = fs.createReadStream(path.join(__dirname, 'problem.txt'))
-exports.solution = fs.createReadStream(path.join(__dirname, 'solution.js'))
+let exercise = require('../../lib/exercise')
+const stdinProcessor = require('../../lib/stdinProcessor')
 
-var input = []; var output = ''
-for (var i = 0; i < 10; i++) {
-  var alien = aliens[Math.floor(Math.random() * aliens.length)]
-  input.push(alien + '\n')
-  output += alien.toUpperCase() + '\n'
-}
+exercise.solution = path.join(__dirname, 'solution.js')
 
-exports.verify = verify({ modeReset: true }, function (args, t) {
-  t.plan(3)
+exercise = stdinProcessor(exercise)
 
-  t.equal(args.length, 1, 'stream-adventure verify YOURFILE.js')
-  var ps = spawn(process.execPath, args)
-  ps.stderr.pipe(process.stderr)
+exercise = comparestdout(exercise)
 
-  ps.stdout.pipe(concat(function (body) {
-    t.equal(body.toString(), output)
-  }))
-
-  ps.on('exit', function (code) {
-    t.equal(code, 0, 'successful exit code')
-  })
-
-  var iv = setInterval(function () {
-    if (input.length) {
-      ps.stdin.write(input.shift())
-    } else {
-      ps.stdin.end()
-      clearInterval(iv)
-    }
-  }, 50)
-})
-
-exports.run = function (args) {
-  var ps = spawn(process.execPath, args)
-  ps.stderr.pipe(process.stderr)
-  ps.stdout.pipe(process.stdout)
-  ps.once('exit', function (code) {
-    if (code) process.exit(code)
-  })
-
-  var iv = setInterval(function () {
-    if (input.length) {
-      ps.stdin.write(input.shift())
-    } else {
-      clearInterval(iv)
-      ps.stdin.end()
-    }
-  }, 50)
-}
+module.exports = exercise
